added group types to global config
