+++ mission/AAW_operation_atlas_NATO_AAF_AAS_Infantry.Altis/cfgBaseSettings.hpp
@@ -20,8 +20,3 @@ markerLocation[] = {
 };
 
 tracks[] = {"AmbientTrack01_F_EPB", "AmbientTrack01_F", "AmbientTrack01a_F", "AmbientTrack01b_F", "AmbientTrack03_F", "AmbientTrack04a_F", "AmbientTrack04_F"};
-
-class GarbageCollector {
-    EnableGarbageCollector = 1;
-    GarbageCollectorTime = 120;
-};
