--- mission/AAW_operation_atlas_NATO_AAF_AAS_Infantry.Altis/cfgRespawn.hpp
@@ -1,3 +1,0 @@
-class CfgRespawn {
-    respawnCountdown = 15; // Time until respawn is available
-};
