+++ mission/AAW_operation_atlas_NATO_AAF_AAS_Infantry.Altis/description.ext
@@ -38,7 +38,7 @@ class CLib {
         "AAW/Logistic",
         "AAW/RespawnUI",
         "AAW/Revive",
-//        "AAW/Supply",
+        "AAW/Supply",
         "AAW/Tickets",
         "AAW/UnitTracker",
         "AAW/VehicleRespawn",
@@ -62,8 +62,6 @@ class AAW {
     #include "cfgSides.hpp"
 
     #include "cfgSectors.hpp"
-
-    #include "cfgRespawn.hpp"
 
     #include "cfgEntities.hpp"
