+++ serverAddons/AAW/addons/AAW_Server/$PBOPREFIX$
@@ -1,1 +1,1 @@
-tc\AAW\addons\AAW_Server
+\tc\AAW\addons\AAW_Server
