+++ serverAddons/AAW/addons/AAW_Server/RespawnUI/cfgRespawn.hpp
@@ -1,0 +1,3 @@
+class CfgRespawn {
+    respawnCountdown = 5; // Time until respawn is available
+};
