+++ serverAddons/AAW/addons/AAW_Server/Squad/fn_clientInit.sqf
@@ -13,7 +13,6 @@
     Returns:
     None
 */
-[QGVAR(GroupTypes), missionConfigFile >> QPREFIX >> "GroupTypes"] call CFUNC(loadSettings);
 
 //GVAR(squadIds) = ("true" configClasses (configFile >> "CfgWorlds" >> "GroupCompany")) apply {getText (_x >> "name")};
 GVAR(squadIds) = [
